{"schema":"decodex/commit/1","summary":"Move scroll capture upward input","authority":"manual"}
